Adapt the website's engine initialisation to parse yaml
